Restructuring routes and views
