+++ package.json
@@ -17,5 +17,8 @@
     "passport-totp": "0.0.2",
     "sprintf": "^0.1.5",
     "thirty-two": "^1.0.2"
+  },
+  "devDependencies": {
+    "nodemon": "^1.19.4"
   }
 }
