+++ routes/api/index.js
@@ -1,0 +1,8 @@
+const ApiRouter = require("express").Router();
+
+ApiRouter.use("/vote", require("./vote"));
+ApiRouter.use("/ballot", require("./ballot"));
+ApiRouter.use("/audit", require("./audit"));
+ApiRouter.use("/account", require("./account"));
+
+module.exports = ApiRouter;
