+++ routes/auth/index.js
@@ -1,0 +1,6 @@
+const AuthRouter = require("express").Router();
+
+AuthRouter.use("/login", require("./login"));
+AuthRouter.use("/logout", require("./logout"));
+
+module.exports = AuthRouter;
