+++ views/vote/vote.ejs
@@ -5,4 +5,4 @@ Name: <%= user.displayName %><br/>
 </p>
 <br>
 <h1>Vote</h1>
-<a href="/logout">Log out</a>
+<a href="/auth/logout">Log out</a>
